moving od docuemnt commands to their own files for rpc
