+++ internal/client/schema.go
@@ -19,19 +19,3 @@ func (c *RpcClient) GetSchema(req rtmv1.QueryWhatIsRequest) (*rtmv1.QueryWhatIsR
 	}
 	return &response, nil
 }
-
-func (c *RpcClient) CreateSchemaDocument(req rtmv1.UploadDocumentRequest) (*rtmv1.UploadDocumentResponse, error) {
-	var response rtmv1.UploadDocumentResponse
-	if err := c.client.Call("Daemon.CreateSchemaDocument", req, &response); err != nil {
-		return nil, err
-	}
-	return &response, nil
-}
-
-func (c *RpcClient) GetSchemaDocument(req rtmv1.GetDocumentRequest) (*rtmv1.GetDocumentResponse, error) {
-	var response rtmv1.GetDocumentResponse
-	if err := c.client.Call("Daemon.GetSchemaDocument", req, &response); err != nil {
-		return nil, err
-	}
-	return &response, nil
-}
